api/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
click==8.3.0
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
joblib==1.5.2
mysqlclient==2.2.6
nltk==3.9.2
openai==2.5.0
pydantic==2.12.3
pydantic_core==2.41.4
regex==2025.9.18
sniffio==1.3.1
sqlparse==0.5.3
textblob==0.19.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
